requirements.txt
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
greenlet==3.3.1
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==16.0
